# PR #4: Dashboard Cleanup & Optimization - Complete Consolidation