--- a/src/context_cleaner/__init__.py
+++ b/src/context_cleaner/__init__.py
@@ -1,8 +1,9 @@
 """
-Context Cleaner - Advanced productivity tracking and context optimization for AI-assisted development.
+Context Cleaner - Advanced productivity tracking and context optimization.
 
-This package provides comprehensive tools for monitoring and improving development productivity
-through intelligent context analysis, performance tracking, and optimization recommendations.
+This package provides comprehensive tools for monitoring and improving development
+productivity through intelligent context analysis, performance tracking, and
+optimization recommendations.
 """
 
 __version__ = "0.2.2"