--- a/src/context_cleaner/optimization/__init__.py
+++ b/src/context_cleaner/optimization/__init__.py
@@ -20,6 +20,7 @@ try:
         HealthLevel,
         UsageInsight,
     )
+
     CACHE_DASHBOARD_AVAILABLE = True
 except ImportError:
     # Graceful fallback when cache module is not available